+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Add `ContainerCreateOptsBuilder::manage_password`
 * Add `ContainerCreateOptsBuilder::mounts`
 * Add `ContainerCreateOptsBuilder::passwd_entry`
+* Add `ImageBuildOptsBuilder::all_platforms`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync